+++ src/components/Navbars/NavBar.jsx
@@ -14,11 +14,7 @@ export default function NavBar() {
           {
             link: "/",
             title: "Home",
-          },
-          {
-            link: "/dash",
-            title: "Dashboard",
-          },
+          },  
         ].map((route) => (
           <A
             className="no-underline text-teal-dark border-b-2 border-teal-dark uppercase tracking-wide font-bold text-xs py-3 mr-8"
